+++ data_fast_insights/resources/literals_mapping.py
@@ -4,5 +4,5 @@ SHOWCASE_LITERALS_MAPPING = {
     'high_perc': lambda kwargs: 'Share of Objects Better than the Total Mean (Segment Quality), %',
     'perc_of_total': lambda kwargs: 'Segment Size, %',
     'target_delta': lambda kwargs: f'{kwargs["target_name"]} Segment Difference from Total Mean, %',
-    'group_importance': lambda  kwargs: 'Group Importance'
+    'target_delta_by_volume': lambda  kwargs: 'Group Effect'
 }